# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateTaskRequest:
    properties:
      endDate:
        description: 结束日期
        example: "2024-12-31"
        type: string
      moduleId:
        description: 关联模块ID
        example: UUID
        type: string
      planHours:
        description: 计划工时
        example: 8
        type: number
      projectId:
        description: 关联项目ID
        example: UUID
        type: string
      startDate:
        description: 开始日期
        example: "2024-01-01"
        type: string
      storyId:
        description: 关联需求ID
        example: UUID
        type: string
      taskRichText:
        description: 任务描述(富文本)
        example: <p>任务描述</p>
        type: string
      taskStatus:
        description: 任务状态
        example: "0"
        type: string
      taskTitle:
        description: 任务标题
        example: 任务标题
        type: string
      taskType:
        description: 任务类型
        example: "0"
        type: string
      userId:
        description: 负责人ID
        example: UUID
        type: string
      versionId:
        description: 关联版本ID
        example: UUID
        type: string
    required:
    - endDate
    - planHours
    - projectId
    - startDate
    - taskTitle
    - userId
    type: object
...
  models.Response:
    properties:
      code:
        description: 状态码 0=成功 -1=失败
        example: 0
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
      message:
        description: 提示信息
        example: ok
        type: string
    type: object
...
  models.TaskResponse:
    properties:
      actualHours:
        description: 实际工时
        example: 6
        type: number
      avatar:
        description: 负责人头像
        example: https://xxx/avatar.jpg
        type: string
      createDate:
        description: 创建时间
        example: "2024-01-01 12:00:00"
        type: string
      creatorId:
        description: 创建人ID
        example: UUID
        type: string
      creatorName:
        description: 创建人姓名
        example: 管理员
        type: string
      endDate:
        description: 结束日期
        example: "2024-12-31"
        type: string
      moduleId:
        description: 关联模块ID
        example: UUID
        type: string
      moduleTitle:
        description: 关联模块标题
        example: 模块名称
        type: string
      percent:
        description: 任务进度百分比
        example: 50
        type: integer
      planHours:
        description: 计划工时
        example: 8
        type: number
      projectId:
        description: 关联项目ID
        example: UUID
        type: string
      projectTitle:
        description: 关联项目标题
        example: crudelis
        type: string
      realName:
        description: 负责人姓名
        example: 张三
        type: string
      startDate:
        description: 开始日期
        example: "2024-01-01"
        type: string
      storyId:
        description: 关联需求ID
        example: UUID
        type: string
      storyTitle:
        description: 关联需求标题
        example: 需求标题
        type: string
      taskId:
        description: 任务ID
        example: UUID
        type: string
      taskNum:
        description: 任务编号
        example: 1
        type: integer
      taskRichText:
        description: 任务描述(富文本)
        example: <p>任务描述</p>
        type: string
      taskStatus:
        description: 任务状态
        example: "0"
        type: string
      taskTitle:
        description: 任务标题
        example: 任务标题
        type: string
      taskType:
        description: 任务类型
        example: "0"
        type: string
      userId:
        description: 负责人ID
        example: UUID
        type: string
      version:
        description: 关联版本号
        example: v1.0.0
        type: string
      versionId:
        description: 关联版本ID
        example: UUID
        type: string
    type: object
...
  models.UpdateTaskFieldRequest:
    type: object
  models.UpdateTaskNextRequest:
    properties:
      changeRichText:
        description: 流转说明(富文本)
        example: <p>流转说明</p>
        type: string
      taskStatus:
        description: 任务状态
        example: "0"
        type: string
    required:
    - taskStatus
    type: object
  models.UpdateTaskRequest:
    properties:
      endDate:
        description: 结束日期
        example: "2024-12-31"
        type: string
      moduleId:
        description: 关联模块ID
        example: UUID
        type: string
      planHours:
        description: 计划工时
        example: 8
        type: number
      projectId:
        description: 关联项目ID
        example: UUID
        type: string
      startDate:
        description: 开始日期
        example: "2024-01-01"
        type: string
      storyId:
        description: 关联需求ID
        example: UUID
        type: string
      taskRichText:
        description: 任务描述(富文本)
        example: <p>任务描述</p>
        type: string
      taskStatus:
        description: 任务状态
        example: "0"
        type: string
      taskTitle:
        description: 任务标题
        example: 任务标题
        type: string
      taskType:
        description: 任务类型
        example: "0"
        type: string
      userId:
        description: 负责人ID
        example: UUID
        type: string
      versionId:
        description: 关联版本ID
        example: UUID
        type: string
    required:
    - endDate
    - planHours
    - projectId
    - startDate
    - taskTitle
    - userId
    type: object
...
  utils.PageResult:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
host: localhost:9191
info:
  contact: {}
  description: Hive Admin Go 后端 API 接口文档
  title: Hive Admin API
  version: "1.0"
paths:
...
  /dev/tasks:
    delete:
      consumes:
      - application/json
      description: 批量删除任务
      parameters:
      - description: 任务ID列表
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除任务
      tags:
      - 开发管理-任务管理
    get:
      consumes:
      - application/json
      description: 分页获取任务列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 任务编号
        in: query
        name: taskNum
        type: integer
      - description: 任务标题
        in: query
        name: taskTitle
        type: string
      - description: 项目ID
        in: query
        name: projectId
        type: string
      - description: 版本ID
        in: query
        name: versionId
        type: string
      - description: 任务状态，支持多选：1,2
        in: query
        name: taskStatus
        type: string
      - description: 需求ID
        in: query
        name: storyId
        type: string
      - description: 排序参数 排序时仅支持：taskTitle、taskStatus、startDate、endDate 的排序
        in: query
        name: sorts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/utils.PageResult'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/models.TaskResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取任务列表
      tags:
      - 开发管理-任务管理
    post:
      consumes:
      - application/json
      description: 创建新任务
      parameters:
      - description: 任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建任务
      tags:
      - 开发管理-任务管理
  /dev/tasks/{taskId}:
    put:
      consumes:
      - application/json
      description: 更新任务信息
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: string
      - description: 任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新任务
      tags:
      - 开发管理-任务管理
  /dev/tasks/{taskId}/field:
    put:
      consumes:
      - application/json
      description: 更新任务的单个字段，仅可修改：userId、taskType、startDate、endDate
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: string
      - description: 字段信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新任务字段
      tags:
      - 开发管理-任务管理
  /dev/tasks/{taskId}/next:
    put:
      consumes:
      - application/json
      description: 更新任务状态
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: string
      - description: 任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskNextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 任务流转状态
      tags:
      - 开发管理-任务管理
  /dev/tasks/{taskNum}:
    get:
      consumes:
      - application/json
      description: 根据任务编号获取任务详情
      parameters:
      - description: 任务编号
        in: path
        name: taskNum
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TaskResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取任务详情
      tags:
      - 开发管理-任务管理
  /dev/tasks/all:
    get:
      consumes:
      - application/json
      description: 获取所有任务（不分页）
      parameters:
      - description: 任务编号
        in: query
        name: taskNum
        type: integer
      - description: 任务标题
        in: query
        name: taskTitle
        type: string
      - description: 项目ID
        in: query
        name: projectId
        type: string
      - description: 版本ID
        in: query
        name: versionId
        type: string
      - description: 任务状态
        in: query
        name: taskStatus
        type: integer
      - description: 需求ID
        in: query
        name: storyId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TaskResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有任务
      tags:
      - 开发管理-任务管理
  /dev/tasks/batch:
    post:
      consumes:
      - application/json
      description: 批量创建任务
      parameters:
      - description: 任务信息列表
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/models.CreateTaskRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 批量创建任务
      tags:
      - 开发管理-任务管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"